Reorganize libs a bit
